tests/wpt/tests/css/css-link-params/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of link-parameters properties</title>
<link rel="help" href="https://drafts.csswg.org/css-link-params-1/#link-param-prop">
<meta name="assert" content="Properties inherit according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_not_inherited('link-parameters', 'none', 'param(--a)')
</script>
</body>
</html>
